move resource view to top and adjust selected units view
